django_app/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-telegram-bot==20.7
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-decouple==3.8
Pillow==10.1.0
django-extensions==3.2.3
django-admin-interface==0.30.1
django-colorfield==0.10.1
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
gspread==5.12.0
oauth2client==4.1.3
dj-database-url==2.1.0

